+++ CHANGELOG.md
@@ -4,10 +4,14 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Added
+- fix for false detection of invisible dog (TC61)
+- fix for missing automatic fault cancelation after last dog that dropped the ball at the gate (manually marked fault) is doing re-run (TC62)
+- fix for false detection of OK crossing due to sensors noise (TC63)
 
 ### Changed
 - MAJOR change in filtering algorithm. Filtering threshold changed from 6ms to 5679us (new magic value). Added N+2 filtering for the same beams (S1 or S2)
 - added TC59 and TC60 related to new filtering algorithm
+- optimized dogs times write to the SD card
 
 ## [1.17.0] - 2024-06-30